rename: lib, api, utils 파일 루트로 이동
