+++ puppy-mate/app/components/SnapShotMap.tsx
@@ -1,9 +1,9 @@
 import { Map, Polyline, MapProps } from 'react-kakao-maps-sdk';
-import { getCenterAndLevel } from '@/app/(map)/utils/getCenterAndLevel';
+import { getCenterAndLevel } from '@/utils/getCenterAndLevel';
 
 export default function SnapShotMap({
   coordinates = [{ lat: 37.566535, lng: 126.977125 }],
-  size = 300,
+  size = 300
 }: {
   coordinates: { lat: number; lng: number }[];
   size: number;
@@ -13,10 +13,10 @@ export default function SnapShotMap({
   return (
     <div
       style={{ width: `${size}px`, height: `${size}px` }}
-      className='overflow-auto mx-auto'
+      className="overflow-auto mx-auto"
     >
       <Map
-        className='w-full h-full bg-gray-100 rounded'
+        className="w-full h-full bg-gray-100 rounded"
         center={center}
         level={level}
         draggable={false}
@@ -27,9 +27,9 @@ export default function SnapShotMap({
         <Polyline
           path={coordinates} // 폴리라인 경로
           strokeWeight={5} // 선 두께
-          strokeColor='#FF0000' // 선 색상
+          strokeColor="#FF0000" // 선 색상
           strokeOpacity={0.8} // 선 투명도
-          strokeStyle='solid' // 선 스타일
+          strokeStyle="solid" // 선 스타일
         />
       </Map>
     </div>
