--- puppy-mate/utils/supabase/server.ts
@@ -1,27 +1,0 @@
-import { createServerClient } from '@supabase/ssr';
-import { cookies } from 'next/headers';
-
-export async function createClient() {
-  const cookieStore = await cookies();
-
-  return createServerClient(
-    process.env.SUPABASE_URL!,
-    process.env.SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        getAll() {
-          return cookieStore.getAll();
-        },
-        setAll(cookieToSet) {
-          try {
-            cookieToSet.forEach(({ name, value, options }) =>
-              cookieStore.set(name, value, options)
-            );
-          } catch (error) {
-            console.error('Error setting cookies:', error);
-          }
-        }
-      }
-    }
-  );
-}
